Init computed season attributes to null
